# micropos/pos-gateway/src/main/java/com/micropos/gateway/dto/ProductDto.java
package com.micropos.gateway.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.Data;

@Data
public class ProductDto {

    @JsonProperty("id")
    private String id;
    @JsonProperty("name")
    private String name;
    @JsonProperty("price")
    private Double price;
    @JsonProperty("image")
    private String image;

}